more use of integer constants
git-svn-id: https://svn.r-project.org/R/trunk@47198 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/base/R/backsolve.R
@@ -35,7 +35,7 @@ backsolve <- function(r, x, k=ncol(r), upper.tri = TRUE, transpose = FALSE)
 	    b  = x, ldb= k,	  nb = nb,
 	    x  = matrix(0, k, nb),
 	    job = job,
-	    info = integer(1),
+	    info = integer(1L),
 	    DUP = FALSE, PACKAGE = "base")[c("x","info")]
     if(z$info != 0)
 	stop(gettextf("singular matrix in 'backsolve'. First zero in diagonal [%d]", z$info), domain = NA)
