+++ src/library/stats/R/kmeans.R
@@ -29,7 +29,7 @@ function(x, centers, iter.max = 10, nstart = 1,
                                 nc = integer(k), double(k), double(k), integer(k),
                                 double(m), integer(k), integer(k),
                                 as.integer(iter.max), wss = double(k),
-                                ifault = as.integer(0), PACKAGE="stats")
+                                ifault = as.integer(0L), PACKAGE="stats")
                        switch(Z$ifault,
                               stop("empty cluster: try a better set of initial centers",
                                    call.=FALSE),
@@ -89,7 +89,7 @@ function(x, centers, iter.max = 10, nstart = 1,
             mm <- nrow(cn)
             if(mm < k)
                 stop("more cluster centers than distinct data points.")
-            centers <- cn[sample(1:mm, k), , drop=FALSE]
+            centers <- cn[sample(1L:mm, k), , drop=FALSE]
         }
     } else {
 	centers <- as.matrix(centers)
@@ -107,7 +107,7 @@ function(x, centers, iter.max = 10, nstart = 1,
     if(nstart >= 2 && !is.null(cn)) {
         best <- sum(Z$wss)
         for(i in 2:nstart) {
-            centers <- cn[sample(1:mm, k), , drop=FALSE]
+            centers <- cn[sample(1L:mm, k), , drop=FALSE]
             ZZ <- do_one(nmeth)
             if((z <- sum(ZZ$wss)) < best) {
                 Z <- ZZ
@@ -116,7 +116,7 @@ function(x, centers, iter.max = 10, nstart = 1,
         }
     }
     centers <- matrix(Z$centers, k)
-    dimnames(centers) <- list(1:k, dimnames(x)[[2]])
+    dimnames(centers) <- list(1L:k, dimnames(x)[[2L]])
     cluster <- Z$c1
     if(!is.null(rn <- rownames(x)))
         names(cluster) <- rn
