+++ src/library/stats/R/na.ts.R
@@ -27,14 +27,14 @@ na.contiguous.default <- function(object, ...)
     if(!sum(good)) stop("all times contain an NA")
     tt <- cumsum(!good)
     ln <- sapply(0:max(tt), function(i) sum(tt==i))
-    seg <- (seq_along(ln)[ln==max(ln)])[1] - 1
+    seg <- (seq_along(ln)[ln==max(ln)])[1L] - 1
     keep <- (tt == seg)
     st <- min(which(keep))
     if(!good[st]) st <- st + 1
     en <- max(which(keep))
-    omit <- integer(0)
+    omit <- integer(0L)
     n <- NROW(object)
-    if(st > 1) omit <- c(omit, 1:(st-1))
+    if(st > 1) omit <- c(omit, 1L:(st-1))
     if(en < n) omit <- c(omit, (en+1):n)
     cl <- class(object)
     if(length(omit)) {
